+++ package.json
@@ -53,10 +53,11 @@
     "@mui/styles": "^5.1.1",
     "@mui/system": "^5.1.1",
     "@mui/utils": "^5.1.1",
-    "@near-wallet-selector/core": "v4.0.0",
-    "@near-wallet-selector/modal-ui": "^4.0.0",
-    "@near-wallet-selector/near-wallet": "v4.0.0",
-    "@near-wallet-selector/sender": "v4.0.0",
+    "@near-wallet-selector/core": "^5.0.1",
+    "@near-wallet-selector/modal-ui": "^5.0.1",
+    "@near-wallet-selector/near-wallet": "^5.0.1",
+    "@near-wallet-selector/my-near-wallet": "^5.0.1",
+    "@near-wallet-selector/sender": "^5.0.1",
     "ethers": "^5.6.4",
     "@react-icons/all-files": "^4.1.0",
     "@reduxjs/toolkit": "^1.6.2",
