+++ src/utils/wallet-selector/wallet-selector-compat.ts
@@ -2,12 +2,13 @@ import { setupWalletSelector } from "@near-wallet-selector/core";
 import type { WalletSelector } from "@near-wallet-selector/core";
 import { setupModal } from "@near-wallet-selector/modal-ui";
 import { setupNearWallet } from "@near-wallet-selector/near-wallet";
+import { setupMyNearWallet } from "@near-wallet-selector/my-near-wallet";
 import { setupSender } from "@near-wallet-selector/sender";
 import * as nearAPI from "near-api-js";
 import BN from "bn.js";
 import { nethIcon, nearWalletIcon, senderWalletIcon } from "./assets/icons";
 
-import { setupMetaMask } from "./metamask/metamask";
+import { setupMetaMask } from "./neth/neth";
 
 import "./modal-ui.css";
 
@@ -77,6 +78,9 @@ export const getSelector = async ({
     debug: true,
     modules: [
       setupNearWallet({
+        iconUrl: nearWalletIcon,
+      }),
+      setupMyNearWallet({
         iconUrl: nearWalletIcon,
       }),
       setupSender({
